workflow/service_chaining/drift_detection/requirements.txt
flask>=1.0.2
google-cloud-storage
data-drift-detector-mightyhive
